Break up CasualtySelector.java, isolate functionalities (notably AA)
